test: move config tests to separate file
